+++ lib/browser/base.rb
@@ -50,16 +50,6 @@ module Browser
       @device ||= Device.new(ua)
     end
 
-    # Return true if browser is modern (Webkit, Firefox 17+, IE9+, Opera 12+).
-    def modern?
-      warn <<~TEXT
-        Browser::Base#modern? is now deprecated.
-        Check https://github.com/fnando/browser/issues/435 for details.
-      TEXT
-
-      Browser.modern_rules.any? {|rule| rule === self } # rubocop:disable Style/CaseEquality
-    end
-
     # Detect if browser is Microsoft Internet Explorer.
     def ie?(expected_version = nil)
       InternetExplorer.new(ua).match? &&